requirements.txt
numpy
pygame==1.9.6
PyOpenGL==3.1.4
PyOpenGL-accelerate==3.1.4
imageio